+++ src/dist_s1/workflows.py
@@ -312,6 +312,7 @@ def run_dist_s1_sas_prep_workflow(
     product_dst_dir: str | Path | None = None,
     bucket: str | None = None,
     bucket_prefix: str = '',
+    n_workers_for_despeckling: int = 5,
 ) -> RunConfigData:
     run_config = run_dist_s1_localization_workflow(
         mgrs_tile_id,
@@ -331,6 +332,7 @@ def run_dist_s1_sas_prep_workflow(
     run_config.product_dst_dir = product_dst_dir
     run_config.bucket = bucket
     run_config.bucket_prefix = bucket_prefix
+    run_config.n_workers_for_despeckling = n_workers_for_despeckling
     return run_config